requirements.txt
# Install requirements for GLaDOS2
discord.py
python-dateutil
requests
pyenchant
beautifulsoup4
PySocks
lxml
nltk
matplotlib
xmltodict
